--- src/browser.js
@@ -1,3 +1,0 @@
-import totallyRandom from "./totallyRandom";
-
-export default totallyRandom;
